ci: Added prefer-single-widget-per-file
