+++ lib/src/widgets/message_boxes.dart/message_box.dart
@@ -1,8 +1,8 @@
-import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 
-import '../../l10n/locale_keys.g.dart';
-import '../models/meta_data_log.dart';
+import '../../models/meta_data_log.dart';
+import 'error_message_box.dart';
+import 'warning_message_box.dart';
 
 /// Colored box widget with an icon, a title and a message.
 ///
@@ -102,34 +102,4 @@ class MessageBox extends StatelessWidget {
         return WarningMessageBox(title: log.title, message: log.message);
     }
   }
-}
-
-/// [MessageBox] which represents an error.
-class ErrorMessageBox extends MessageBox {
-  /// Creates a [ErrorMessageBox].
-  ErrorMessageBox({
-    super.key,
-    required super.title,
-    required super.message,
-  }) : super(
-          textColor: const Color.fromARGB(255, 255, 82, 82),
-          backgroundColor: const Color.fromARGB(255, 253, 234, 236),
-          iconData: Icons.error,
-          titleTag: LocaleKeys.message_box_title_error.tr(),
-        );
-}
-
-/// [MessageBox] which represents a warning.
-class WarningMessageBox extends MessageBox {
-  /// Creates a [WarningMessageBox].
-  WarningMessageBox({
-    super.key,
-    required super.title,
-    required super.message,
-  }) : super(
-          textColor: const Color.fromARGB(255, 255, 145, 0),
-          backgroundColor: const Color.fromARGB(255, 253, 241, 229),
-          iconData: Icons.warning,
-          titleTag: LocaleKeys.message_box_title_warning.tr(),
-        );
 }
